Add elevator control to IntakeCoral.
